Added Spell Checking
Added a CSpell configuration for spell-checking the contents of the
repository, checked all files, and corrected any spelling mistakes. The
spell-checking was integrated into the GitHub Actions workflow, so that
the spelling is checked on every push and pull request to the master
branch. The documentation still needs to be updated to reflect the new
spell-checking process.

Furthermore, removed LaTeX commands for accented characters from the
bibliography file, as they we are using Pybtex to handle the
bibliography and it has full Unicode support.