Removed the LICENSE file
The project is dual-licensed under the GNU General Public License
Version 3 (GPL-3.0) or later, and the GNU Lesser General Public License
Version 3 (LGPL-3.0) or later. The GPL-3.0 license is in the COPYING
file and the LGPL-3.0 license is in the COPYING.LESSER file.
Additionally, there is LICENSE file, which contains a note about the
dual-licensing. This is, however, confusing, as GitHub does not
recognize that the file is only a note about the dual-licensing and not
the actual license. This commit removes the LICENSE file and adds a note
about the dual-licensing to the read me. Also, the read me was cleaned
up a bit to fix some of the problems that Markdownlint found.