Added a CITATION.cff
Added a CITATION.cff file, which contains the necessary information to
cite this repository. This file is based on the Citation File Format
(CFF) standard. This file is supported by GitHub and results in a "Cite
this repository" button on the website, which allows users to directly
generate a proper citation for the repository in multiple different
formats.

Also added a bullet point to the changelog about the removal of the
LICENSE file, which I had forgotten to do in the last commit.