Converted the Project to a uv project
Converted the CoRelAy project from a setup.py project to a uv project.
For this, a pyproject.toml file was created, which is configured to do
the same as the setup.py file. The source code was moved from
"src/corelay" to "source/corelay".

The tox configuration was updated and now uses tox-uv to run all
commands via uv instead of directly creating environments. This means,
that all Python environments can now be run without having to install
multiple Python versions. The configuration was also cleaned up.

Support for Python 3.7 was removed, not only because it has already
reached its end-of-life, but also because some of the dependencies
(especially tox-uv) do not support it anymore. The two remaining
supported Python versions (3.8 and 3.9) are now recorded in the
".python-versions" file, which makes it trivial to install them using
uv.

The GitHub Actions workflow that runs the unit tests, linters, and
builds the documentation was updated to point to the new locations of
the source code, unit tests, and configuration files. Also, it now uses
uv to run the tests and linters. Split up the GitHub Actions workflow
job that ran PyLint and Flake8 into two separate jobs, which allows for
better parallelization and faster execution of the workflow. the
"actions/checkout" action used in the GitHub Actions workflow was
updated from version 2 to version 4. The GitHub Actions workflow matrix
configurations for Python 3.7 was removed, as it is no longer supported
by the project. Finally, the workflow was cleaned up and documented.

The usage of uv still needs to be documented, but this will be done
later when the entire documentation will be updated.

Furthermore, the following changes were made:

- The unit tests were moved from the "tests" folder to
  "tests/unit_tests", which clears up space for other test files.
- The tox configuration was moved from the root directory to the "tests"
  directory, which is more appropriate, as tox is mostly used for
  testing.
- The configurations for the linters PyLint and Flake8 were moved from
  the root directory of the repository (in the case of PyLint) and from
  the tox configuration file (in the case of Flake8) into the
  "tests/linters" directory. The CSpell configuration was also moved
  there.