Updated the Dependencies of the Project
Updated the Python dependencies in the "pyproject.toml" to their
respective latest versions. Some of the dependencies no longer support
Python 3.8 and Python 3.9 (as well as Python 3.10). For this reason, the
project was migrated to support Python 3.11, 3.12, and 3.13. The
".python-versions" file, the tox configuration, and the GitHub Actions
workflow were updated to reflect this change. One of the dependencies
("metrohash-python") requires a C++ compiler to build and uses the "c++"
command, which may not be available on all systems. To ensure that users
are not confused by this, a note was added to the README file
explaining that the "c++" command is required to build the project and
showing how to install it on Fedora (one of the systems that do not have
the "c++" command installed by default).)

The configuration for Read the Docs was updated to use the latest
available versions of Ubuntu (24.04) and Python (3.12). It was also
documented.

Furthermore, the following changes were made:

- The `corelay/version.py` file was deleted as it is automatically
  generated during the build process and should not be checked into
  source control.
- The configuration for the GitLab CI, which was stored in the
  ".gitlab-ci.yml" file, was removed. The project is no longer being
  hosted on GitLab, and the CI configuration is no longer needed.