Updated the Documentation
Updated the documentation to reflect the changes made to CoRelAy. First
of all, the Sphinx configuration was updated:

- The "sphinxcontrib.datatemplates" extension was removed, because it
  was not used in the documentation.
- Added the "sphinx.ext.intersphinx" extension to add links in the
  documentation to external documentations, e.g., the Python standard
  library documentation and the NumPy documentation.

The documentation was largely extended and a bit reorganized. The API
reference documentation was moved from the "reference" directory to the
"api-reference" directory. This was done to better reflect the purpose
of the documentation, since there are also bibliographical references in
the documentation. The "Getting Started" section was expanded to new
section with multiple sub-sections: "Installation", "Basic Usage", and
an "Example Project". The installation and basic usage sections were
moved from the original "Getting Started" section. The "Example Project"
section contains a more elaborate example of how to use CoRelAy to
analyze a dataset generated by Zennit using the SpRAy workflow that can
be visualized using ViRelAy. A new "Contributor's Guide" section was
added with sub-sections on how to report issues and feature requests,
and how to contribute code or documentation. Finally, a new "Migration
Guide" section was added to help users migrate from CoRelAy v0.2 to
CoRelAy v0.3. This was done, because the changes made in CoRelAy v0.3
are quite extensive and it is not easy to find out what has changed and
how to adapt existing code to the new version. Finally, more citations
were added to the documentation to relevant literature.

A custom CSS file was added to the documentation to change the alignment
of the documentation text to be justified and to automatically break
long words. This was done to improve the readability of the
documentation and to align it with usual scientific documents. Also, the
width of the labels in the bibliography was set to a fixed size, so that
the labels all have the same width. This was done, because they looked
inconsistent and messy before.

The docstrings in the CoRelAy source code were updated to now use the
":py:*:" directives to link references to classes, functions, methods,
modules, and values to their respective external documentation. The
usage of type hints had to be changed in some places to accommodate the
way that Sphinx AutoDoc and Intersphinx reference types. Also, some of
the docstrings were updated to fix typos, improve clarity, or to add
missing information.

By default, Sphinx AutoDoc only documents the "__init__" method and
leaves out all other special "dunder" methods, such as "__repr__". A
list of the special methods that are used in the CoRelAy source code
was added to the list of methods covered by Sphinx AutoDoc, so that
these methods are also documented.

Usage of the "Literal" type hint was totally removed from the source
code, because it is not properly supported by Sphinx AutoDoc and made
more problems than it solved. The code for retrieving the source code
location of type aliases to "Literal" type hints was removed from the
documentation configuration Python file, as it is no longer needed.

The invocation of Sphinx in tox to build the documentation was updated
to include the "--fresh-env" option, so that the documentation is always
built in a fresh environment. This solves some issues with the
documentation build, where the documentation was not updated correctly
after changes were made to the source code or the documentation itself.

Besides the updates to the documentation, the read me of the repository
was also updated. The introduction was extended to better explain the
CoRelAy library. A section on the features of CoRelAy was added. The
installation instructions and the usage section were now wrapped in a
new "Getting Started" section. The example code in the usage section was
updated to reflect the changes made to the example scripts. The
"Contributing" section was also updated to include a proper description
and links to the contributors guide in the documentation.

Also, the example scripts were updated to reflect the coding style and
docstring conventions used in the CoRelAy library. They were moved from
the "example" directory to the "docs/examples" directory. This was done
because the examples are part part of the documentation and to keep the
root directory clean and free of unnecessary clutter.

Some of the "__repr__" methods in the CoRelAy source code had to be
updated, because Sphinx AutoDoc uses the "__repr__" method to generate
type information for parameters, which was interfering with the usage of
"Annotated" type hints: since the instance of "Param" or "Task" is a
parameter of the "Annotated" type hint, AutoDoc would call the
"__repr__" method on them and interpret the output as the type. This is
obviously weird and not what we want, so the "__repr__" methods were
updated to return a string that is more suitable for the documentation.
Also, the code for retrieving the source code of a lambda expression,
which was used in the "__repr__" methods was massively updated to fix
some edge cases that were not handled correctly before.