Increased the Test Coverage to 100%
Added new unit tests and updated existing ones to increase the test
coverage of the CoRelAy codebase to 100%.

The dependencies of the CoRelAy package and of the Node.js-based linters
were updated to their latest versions, since this is the last commit
before the release. This was done to ensure that everything is
up-to-date and works as expected. The UMAP dependency is now installed
directly from the GitHub repository, since the latest version of UMAP is
not yet available on PyPI, although already being available since
February 2025. The new version fixes some deprecation warnings of
third-party libraries, which are used by CoRelAy, thus ensuring that
CoRelAy does not raise any deprecation warnings itself. Pybtex, a
dependency of the BibTeX extension for Sphinx, which is used to generate
the bibliography in the documentation also causes a deprecation warning
when building the documentation as it is still using the deprecated
"pkg_resources" module for loading plugins. The version of Pybtex
available on PyPI is 0.24.0, which is still using the "pkg_resources"
module. The problem has already been fixed in the development version,
but it is not yet released. Therefore, the development version is
installed directly from the Git repository. If and when a successor to
0.24.0 is released, this dependency can be removed again. The optional
dependencies of the CoRelAy package are now also included in the
"testing" and "linting" dependency groups, so that CoRelAy can be tested
and linted with the optional dependencies installed.

Some minor changes were made to the codebase:

- The "Shaper" flow processor was extended to support dictionaries and
  string indices. It seems like, these features were already expected to
  be present, but they were prevented by some minor mistakes:
  Dictionaries were not supported, because the "Shaper" processor tested
  for the type of the input data by testing if they implemented the
  "Sequence" protocol, which is not the case for dictionaries. Instead,
  dictionaries implement the "Mapping" protocol, which is now also
  tested for. String indices were not supported, because the indices
  were tested for implementing the "Sequence" protocol, which is also
  the case for strings. This meant, that each character of a string was
  treated as a separate index, which is not the intended behavior. Now,
  it is also checked if the indices are not strings.
- The stack level of the warning, which is raised when the old "Slot"
  syntax is used was increased to 2, so that the user can see the
  location of the code that is using the old syntax, instead of seeing
  the location of the warning itself.
- The "get_lambda_expression_source_code" function was updated to fixed
  some use-cases in which the source code of the lambda expression was
  not correctly retrieved.
- The function types that "Slot" and "Plug" support has been expanded to
  include some types that were previously not supported.
- The "get_fully_qualified_name" function was also updated to support
  these new function types.

Some further code changes were made to make the code more testable. This
only includes changes that remove checking code that is never executed.
For example, in the "MetaTracker" meta class, the check if the class has
an attribute "__bases__" was removed, since this attribute is always
present in Python classes.

Also, some minor adjustments were made to the docstrings of the codebase
to include more information or correct the information already present.

Finally, some bugs in the codebase were fixed, that were discovered
during the creation of the new unit tests:

- Fixed a bug in the "RadialBasisFunction" affinity processor: The
  formula for the Radial Basis Function (RBF) kernel was incorrect. The
  distance matrix was not squared resulting in the wrong formula
  K(d) = exp(-d/(2sigma^ 2)) instead of the correct formula
  K(d) = exp(-(d^2)/(2sigma^2))$. This was fixed by squaring the
  distance matrix.
- Fixed a bug in the "Histogram" processor: The processor was using the
  "numpy.histogramdd" function, which computes a multivariate histogram,
  but the processor was meant to compute a histogram over the channels
  of the input data, for which the "numpy.histogramdd" function is not
  suitable. Instead, the "numpy.histogram" function is now used in
  conjunction with the "numpy.stack" function to compute the histogram
  over the channels of the input data. Also, the "Histogram" processor
  was not able to deal with channel-last data, which is now supported.
  The "Histogram" processor created in the "virelay_analysis.py" example
  script was also not working as expected: Although it was using the
  "numpy.histogram" function, it did not account for its return type,
  which is a tuple containing the histogram and the bin edges. Since we
  now have a working implementation of the "Histogram" processor, the
  example script was updated to use it.
- The root directory specified in the CSpell script was not correct,
  which meant that not all files were checked for spelling errors. This
  was fixed by updating the root directory to the correct one.