Removed Node.js From the Deployment Workflow
The GitHub Actions Workflow for the deployment of CoRelAy to PyPI was
copied over from ViRelAy and still contained the installation of
Node.js, which is needed in ViRelAy for the building of the frontend,
but is not required for the building of CoRelAy. For this reason, the
installation of Node.js was removed from the GitHub Actions Workflow for
the deployment.